+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- `check-consul-quorum.rb`: Check how many servers can be lost while maintaining quorum (@roboticcheese)
+- `check-consul-stale-peers.rb`: Check for stale peers in the raft configuration (@roboticcheese)
 
 ## [2.0.1] - 2018-03-27
 ### Security